cleanup of plugin code
