+++ plugins/talk-plugin-profile-data/server/constants.js
@@ -1,0 +1,3 @@
+module.exports = {
+  DOWNLOAD_LINK_SUBJECT: 'download_link',
+};
