+++ plugins/talk-plugin-profile-data/server/typeDefs.js
@@ -1,0 +1,7 @@
+const path = require('path');
+const fs = require('fs');
+
+module.exports = fs.readFileSync(
+  path.join(__dirname, 'typeDefs.graphql'),
+  'utf8'
+);
